refactor domsat tests to junit
