+++ src/test/java/integration/lrgs/domsatrecv/TestDomsat.java
@@ -2,7 +2,8 @@ package integration.lrgs.domsatrecv;
 
 import java.util.logging.Logger;
 
-import lrgs.domsatrecv.DomsatFranklin;
+import lrgs.domsatrecv.DomsatDpc;
+import lrgs.lrgsmain.LrgsConfig;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
@@ -11,16 +12,20 @@ import org.junit.jupiter.api.RepeatedTest;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-@Disabled("Requires domsat.win on java.library.path")
-final class DomsatFranklinTest extends Thread
+@Disabled("Requires DOMSAT DPC installed on localhost")
+final class TestDomsat
 {
-	private static final Logger LOGGER = Logger.getLogger(DomsatFranklinTest.class.getName());
-	private static DomsatFranklin ds;
+	private static final Logger LOGGER = Logger.getLogger(TestDomsat.class.getName());
+	private static DomsatDpc ds;
 
 	@BeforeAll
 	public static void setup()
 	{
-		ds = new DomsatFranklin();
+		String host = System.getProperty("domsat.dpc.test.host", "localhost");
+		int port = Integer.getInteger("domsat.dpc.test.post", 9000);
+		LrgsConfig.instance().dpcHost = host;
+		LrgsConfig.instance().dpcPort = port;
+		ds = new DomsatDpc();
 		ds.init();
 		assertTrue(ds.setEnabled(true), () -> "Enable failed -- see log messages.");
 	}
@@ -32,13 +37,19 @@ final class DomsatFranklinTest extends Thread
 	}
 
 	@RepeatedTest(value = 20, name = RepeatedTest.LONG_DISPLAY_NAME)
-	void domsatFranklin()
+	void dpcDomsat() throws Exception
 	{
-		byte[] packet = new byte[1024];
+		byte packet[] = new byte[1024];
 		int len = ds.getPacket(packet);
+
 		assertNotEquals(0, len, () -> "Timeout");
 		assertNotEquals(-1, len, () -> "Recoverable Error: " + ds.getErrorMsg());
 		assertNotEquals(-2, len, () -> "Fatal Error: " + ds.getErrorMsg());
+		while(len == -3)
+		{
+			Thread.sleep(100L);
+			len = ds.getPacket(packet);
+		}
 		LOGGER.info("Frame received, len=" + len);
 		LOGGER.info(" "
 				+ Integer.toHexString((int) packet[0] & 0xff) + ", "
